Added code structure for QUndoStack
Added a QUndoStack to the document.
Enabling of undo & redo is via signals from the QUndoStack.
Reverting the document will come from winding back the stack
The stack is marked as clean when the document is saved although
the content still exists and can still be undone.
All changes to the document should now be done through QUndoCommand
derived classes.
Changed some method parameters to const references.