Implemented add and remove stitches in StitchQueue
Moved the code to add and remove stitches to the StitchQueue from the
document to the StitchQueue class. This affected the floss usage
functionality so changed this to a function that will recalculate the
usage before it is used for anything.  Consequently the code to update
the floss usage in individual functions has been removed.