Fixed bugs in FilePropertiesDlg and ImportImageDlg
Width and Height boxes had a minimum value of the width and
height of the pattern extents and would not be modified when
the units where changed resulting in values below the minimum.

The size of the pattern was not displayed on the import dialog.
The position of the pattern size was relocated in the ui file
layout to improve the appearance.