Encapsulated palette and stitch data in parent class
It makes sense to treat the pattern stitch data and the palette data
as a single entity so that it can be passed around various functions.
Although the DocumentPalette and StitchData classes are not combined
they are used as members of a Pattern class.
All other files have been updated as they now access the stitch and
palette data via the Pattern class.