Added cut, copy and paste functionality
Action slots moved from MainWindow to Editor.
Cut copy and paste functions moved from StitchData to Pattern.
Pattern now takes a Document parameter in the constructer, defaults to
null if not supplied.  Access function added to get the Document
pointer.
Functions added to StitchData to support cut, copy and paste.
StitchData streaming code updated to stream Backstitch and Knot data.
Renderer now takes a pointer to a Pattern and not a Document.  Allows
the paste pattern to be rendered using the Editor renderer.  Also
required an offset to render it to the mouse position, defaults to
QPoint(0,0) if not supplied.
Functions added to Editor to handle mouse events for the paste function.
Function added to Editor to provide a list of Stitch::Type relative to
the selected stitch type.
Function added to DocumentPalette to add a floss by color.
QUndoCommand derived classes added for cut and paste functions.
MainWindow sets the selection masks directly as setting the actions
didn't trigger the action when setting to the same as its current
selection state.