Implemented text tool
Removed text tool slot from MainWindow as this is incorporated into the
Editor tool modes.
Added tool initialisation call pointers for each tool to the Editor,
this will allow any setup operations required for any of the tools when
they are selected.
Modified the paste slot to just create a paste pattern.  A new generic
function handles initialisation of the paste mode.
Added function to Editor to display the Text dialog and create the
required pattern that will be passed to the paste function.
Modified the mouse release event for paste to use a descriptice text
string to identify the source of the paste operation, this is used in
the undo/redo description.
Refactored the TextToolDlg code which is now simpler.