Added keyboard input for pasting
When using paste based tools including the text tool, the keyboard
may be used to position the pasted pattern after using Ctrl-V to
activate the paste command or after completion of the tool dialog.
Arrow keys are used to move it, enter to finalise the position or escape
to cancel the operation.