Implement filled polygon tool
Add functionality to draw polygon boundary which is filled when the
first cell is reselected.  Keyboard interface allows deletion of last
cell selected or escape to cancel the operation.
Introduced a series of toolCleanup functions to clean up when tools are
deselected, e.g. the polygon tool may have a partial polygon boundary on
screen when another tool is selected.
Changed name references of Polyline to Polygon.
Renamed the icon to polygon.png