Fix bug with reading multiple versions of a scheme
When a scheme is calibrated and the original is write protected because
it is a system version, a replacement is created in the users home
folder, however when restarting kxstitch, both versions are being read.

This fix assumes that the schemes in the users folder are available
first with the system ones last.  When reading schemes, if one with the
same name exists, the one being read will be deleted and not added to
the list of schemes available.