Fix bug in palette view causing crash
When the palette was empty, hovering the mouse over it tries to show a
tooltip but it tried to calculated the position of the floss color cell
first resulting in a divide by zero. This moves the cell calculation inside
the block protected by a check for an empty palette.