Fix several bugs in the color calibration
Correctly set the selected scheme in the list.
Correctly set the initial current color in the color list when changing
the scheme selected.
Change handling of the Ok button press. The color change map is populated
with the name of the scheme and a map of names and colors for any changed
colors. The scheme name will be inserted in the change map when switching
schemes but there may not be any changed colors associated with it.
Check for an empty changed color map before writing the scheme.
The pointer to the current color is reset to null when changing schemes
as the item pointed at gets deleted when the color list is cleared.