Modify handling of file names when writing schemes
Writing schemes would use the scheme name rather than the original file
name.  The original scheme path is stored in the FlossScheme class.
Writing the scheme checks if the original path is writable before trying
to create a writable path. The same file name is always used regardless
of the path.
Include a fix for writing the J&P Coats scheme which requires & to be
replaced with &amp;