Fix bug in pattern renderer.
The printer configuration element for patterns allows the selection of
an area of the pattern to be rendered in the element.  The renderer was
taking account of the width and height of the selection but not the top
left position. This fixes the bug and correctly renders the selected
area.