Implement improved file handling.
Implement exception generation and handling for read and write functions
of the document. Saving the document now generates a backup file. Saving
the document now uses KSaveFile class to implement saving and then
overwriting the file to avoid corrupting the file on a crash or some
other failure to write to the file.