Modify PaletteManagerDlg to use a complete palette
Originally the PaletteManagerDlg used elements of the DocumentPalette to
operate and created a change command for every DocumentFloss that was
modified. This also duplicated some code used in the DocumentPalette to
generate the next available index and symbol. This update allow the
dialog to operate on a complete DocumentPalette and modify it
directly. If commited the new DocumentPalette is swapped with the one in
the Document. Using undo will swap it back for the original one saved.