Fix for open recent file crash
In the fileOpen function the url was added to the file_open_recent
action to maintain the recent files list and the document url was then
set from the url opened, however the url being a reference becomes
corrupted when added to the action and consequently caused a crash when
setting the documents url.
This fix changes the order to set the document url before adding it to
the recent files action.