Fix for bug #22 - Library pattern properties change
This bug failed to save changes of the key code and modifiers to the
library file due to a missing destructor. Additional changes now allow
deletion of the key code by using the delete or backspace key.

Item 3 in the bug report was not able to be replicated. Creating library
patterns with the same key associated resulted in the first found
library pattern used for the key press. No crash occurred.

This update also fixes a layout fault after updating the properties that
rendered the icon to large for the size set by the zoom factor.