Fix for config dialog default scheme and symbols
The config dialog was not storing the text values of the KComboBox
elements used to pick the default scheme and default symbol library.
KConfigXT only stores the index. As the list is dynamically created,
this may result in the incorrect selection being made. This fix now
stores the text value itself which maintains consistency.