Added config dialog options for default stitch and backstitch strands
The configuration file has options for the default number of strands for
stitches and backstitches, but the configuration dialog had no options
for changing the values used. Two combo boxes have been added to the
palette configuration to allow setting this value in the configuration
file.

GUI: Two new strings have been added in PaletteConfigPage.ui