Fix for crashes deferencing null pointer from QWidget::paintEngine()
QWidget::paintEngine may no longer return a valid pointer, so dereferencing
it caused a crash. Test for null pointer and if found, assume a QWidget
is being drawn on. This code is designed to draw a reference border
around an element on screen only. There should be a better way to do this.