Fix rubberband selections in various dialogs
Bug 388454 identified a problem creating elements in the print setup
dialog. Investigation shows there appears to be a Qt bug in QWidget
derived classes when used in QDialogs, which blocks mouse events when a
mouse button is pressed and a QToolTip is displayed.

A minimal application that replicates this is in github:
https://github.com/SAllewell/tooltip-bug

The original code worked in Qt 5.6, but fails in Qt 5.10
The version that introduced the regression is unknown.

It has been logged as QTBUG-65599
BUG: 388454
FIXED-IN: 2.1.0