Fix for importing images for V6 of ImageMagick
The getPixelColor in V6 of ImageMagick does not appear to return the
same information as in V7, consequently importing images has resulted in
a black image when using V6 of ImageMagick.  This fix reverts the change
made in commit 295773f44bfda1227d85edf065a8de14dc889159 when using V6.

Big thanks to Sean Enck for reporting and helping diagnose the problem.