Fix file opening bug
A change in KDE/Qt has resulted in a QTemporaryFile size() of 0 after
KIO::copy_file overwrites it, previous versions show the size of the
copied file.  As such garbage was read instead of the file content.

Fixed by closing the temporary file after opening it, reopening it using
a QFile which then has the correct size.
BUG: 407875
FIXED-IN: 2.2.0