Switch from XKeyLock to KModifierKeyInfo
Use KModifierKeyInfo from the KGuiAddons framework to get the status
of keys. In particular, use it to get the status of the Caps Lock key
instead of the local XKeyLock implementation.

The KModifierKeyInfo object is kept as class member to avoid recreating
it at each key press in Alphabet mode.