Manage page sizes using QPageSize in the printing configuration
Instead of specifying statically in the UI all the paper sizes, use the
existing helper to get available sizes, and fill the combobox using
QPageSize.

This way the list of sizes for printing is no more duplicated in this UI
as well, and there are no customer translations for the paper sizes.